Update to use doubles

+++ src/main/java/ca/bcit/comp2522/termproject/jaguarundi/Interactable.java
@@ -1,8 +1,8 @@
 package ca.bcit.comp2522.termproject.jaguarundi;
 
 public abstract class Interactable {
-    public abstract int getXPosition();
-    public abstract int getYPosition();
-    public abstract int getWidth();
-    public abstract int getHeight();
+    public abstract double getXPosition();
+    public abstract double getYPosition();
+    public abstract double getWidth();
+    public abstract double getHeight();
 }
